refactor(pi-sequential-thinking): extract config resolution into config.ts
Move ~270 lines of config-precedence logic out of index.ts:

- ConfigSource, SeqThinkConfig, SeqThinkConfigWithSources types
- normalizeString, normalizeNumber, splitParams, resolveEffectiveLimits
- resolveConfigPath, parseConfig, loadSettingsConfig
- loadConfigFileWithSources, mergeConfigWithSources, loadConfigWithSources
- warnIgnoredLegacyConfigFiles, resolveEffectiveConfig, getHomeDir

index.ts re-exports the previously inline helpers via `export ... from
"./config.js"`, so existing tests importing them from index.ts keep
working unchanged. No behavior changes — config precedence, source
labels, and error messages are identical.

index.ts: 1036 → 614 LoC.
